extract logic to a component
